refactor: extract FsUtils.ts
