+++ src/features/PathConfiguration.ts
@@ -37,10 +37,13 @@ interface PathConfigurationValues {
 }
 
 export default class PathConfiguration {
+    static configuration = new PathConfiguration();
+
     readonly data: PathConfigurationValues;
 
-    constructor() {
+    private constructor() {
         this.data = {};
+        this.update();
     }
 
     update(fileUri?: vs.Uri) {
@@ -61,7 +64,8 @@ export default class PathConfiguration {
         this.data.enableFolderTrailingSlash = codeConfiguration.get('enableFolderTrailingSlash');
         this.data.ignoredFilesPattern = codeConfiguration.get('ignoredFilesPattern');
         this.data.ignoredPrefixes = codeConfiguration.get('ignoredPrefixes');
-        this.data.homeDirectory = process.env[process.platform === 'win32' ? 'USERPROFILE' : 'HOME'];
+        this.data.homeDirectory =
+            process.env[process.platform === 'win32' ? 'USERPROFILE' : 'HOME'];
 
         const workspaceRootFolder = vs.workspace.workspaceFolders
             ? vs.workspace.workspaceFolders[0]
